reorganize graph stuff to graph widget
